+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## 30.0.7
+New:
+- `GeoMultiSelect`: Now pagination is applied on grouped items not on groups
 ## 30.0.6
 Fix:
 - `GeoChart`: Fix negative width for axis ticks